requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Markdown==3.5.1
Pillow==10.1.0
email-validator==2.1.0.post1
python-slugify==8.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
python-slugify
flask-admin
wtforms_sqlalchemy